requirements.txt
appdirs
astroid
black
certifi
chardet
click
cycler
Cython
datasets 1.9.0
dill
einops
filelock
fire
flake8
GPUtil
idna
isort
joblib
jsonlines
kiwisolver
lazy-object-proxy
matplotlib
mccabe
multiprocess
mypy-extensions
numpy
opencv-python
packaging
pandas
pathspec
pep8
Pillow
pyaml
pyarrow
pycocotools
pycodestyle
pyflakes
pylint
pyparsing
python-dateutil
pytz
PyYAML
regex
requests
sacremoses
scipy
six
termcolor
timm
tokenizers
toml
torch
torchvision
tqdm
transformers 4.8.2
typed-ast
typing-extensions
urllib3
wget
wrapt
xxhash
sphinx
sphinx_rtd_theme
recommonmark

